Actualización de Servicios y lógica de negocio.

+++ todo-deportes/src/main/java/com/tododeportes/backend/ngc/UnidadDeportivaNGC.java
@@ -20,6 +20,6 @@ public interface UnidadDeportivaNGC {
 	
 	public TbUnidadesDeportivas obtenerUnidadDeportivaxId(int id)  throws ExcepcionesNGC;
 		
-	public List<TbCanchasxunidadDeportiva> listarCanchaxUnidadDeportiva(Long idUnidadDeportiva) throws ExcepcionesNGC;
+	public List<TbCanchasxunidadDeportiva> listarCanchaxUnidadDeportiva(int idUnidadDeportiva) throws ExcepcionesNGC;
 	
 }
